Non-Lighting stuff...
Added functionality to display a pixel map on the faces of
parallelograms.
Added framework for a polygonal wireframe renderer that will be
thousands of times faster than the ray tracing renderer.
Attempted to fix that pesky lighting bug, to no avail.
Updated a lot of documentation and syntax, specifically if statements
without braces in order to conform to new clenliness standards imposed
by netbeans.